+++ src/main/java/com/wex/exchangetransactions/dto/TransactionRetrieveResponseDTO.java
@@ -2,6 +2,7 @@ package com.wex.exchangetransactions.dto;
 
 import com.wex.exchangetransactions.annotation.RoundFractionalValue;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import static com.wex.exchangetransactions.exception.error.ValidationErrorMessages.RETRIEVE_CONVERTED_AMOUNT_ROUNDED_MESSAGE;
@@ -10,14 +11,14 @@ public record TransactionRetrieveResponseDTO(
         String transactionId,
         String description,
         LocalDate transactionDate,
-        Double originalAmount,
+        BigDecimal originalAmount,
         Double exchangeRate,
         @RoundFractionalValue(fractionDigits = 2, message = RETRIEVE_CONVERTED_AMOUNT_ROUNDED_MESSAGE)
-        Double convertedAmount) {
+        BigDecimal convertedAmount) {
     public static TransactionRetrieveResponseDTO buildResponseDTO(
             TransactionResponseDTO purchaseTransaction,
             Double exchangeRate,
-            Double convertedAmount
+            BigDecimal convertedAmount
     ){
         return new TransactionRetrieveResponseDTO(
                 purchaseTransaction.id(),
